moved config classes to valid OASP packages (note: we could also move them to toplevel package where SpringBootApp is located to have them together)
