+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/general/service/impl/config/BaseWebSecurityConfig.java
@@ -1,4 +1,4 @@
-package io.oasp.gastronomy.restaurant.general.configuration;
+package io.oasp.gastronomy.restaurant.general.service.impl.config;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
