+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/general/service/impl/config/ServletInitializer.java
@@ -1,4 +1,4 @@
-package io.oasp.gastronomy.restaurant.general.configuration;
+package io.oasp.gastronomy.restaurant.general.service.impl.config;
 
 import io.oasp.gastronomy.restaurant.SpringBootApp;
