+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/general/service/impl/config/TestWebSecurityConfig.java
@@ -1,4 +1,4 @@
-package io.oasp.gastronomy.restaurant.general.configuration;
+package io.oasp.gastronomy.restaurant.general.service.impl.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,6 +11,7 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.authentication.www.BasicAuthenticationEntryPoint;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 
+import io.oasp.gastronomy.restaurant.general.service.impl.config.BaseWebSecurityConfig;
 import io.oasp.module.basic.common.api.config.SpringProfileConstants;
 
 /**
